requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.13
openai>=1.3.0
transformers>=4.35.0
pinecone>=2.2.4
torch>=2.0.0